main-project-web/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
cycler==0.10.0
Django==3.2.9
easyocr==1.4.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.9.0
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
Pillow==8.2.0
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.26.0
rfc3986==1.5.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
tifffile==2021.11.2
torch==1.10.0
torchvision==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.7
